refactor: move Propagator class to new propagator package
