+++ examples/example.py
@@ -42,7 +42,7 @@ sat = gp.Spacecraft.from_dict(sat_params)
 fm = gp.ForceModel(name='LowEarthProp_ForceModel', point_masses=['Luna', 'Sun'], drag=gp.ForceModel.DragForce(),
                    srp=True, gravity_field=gp.ForceModel.GravityField(degree=10, order=10))
 prop = gp.PropSetup('LowEarthProp', fm=fm, accuracy=9.999999999999999e-12,
-                    gator=gp.PropSetup.Propagator(name='LowEarthProp', integrator='RungeKutta89'))
+                    gator=gp.Propagator(name='LowEarthProp', integrator='RungeKutta89'))
 
 # Build an ImpulsiveBurn that will be used in the Maneuver command
 toi = gp.ImpulsiveBurn('IB1', sat.GetCoordinateSystem(), [0.2, 0, 0])
