+++ gmatpyplus/__init__.py
@@ -56,6 +56,7 @@ from .hardware import *
 from .interpreter import *
 from .navigation import *
 from .parameter import *
+from .propagator import *
 from .solar_system import *
 from .solver import *
 from .spacecraft import *
